requirements.txt
aiohttp==3.9.1
python-dotenv==1.0.0
faiss-cpu==1.7.4
scikit-learn==1.3.2
numpy==1.24.3 